starting to get some semblance of a working 3d animated mesh system

+++ saguinus.h
@@ -135,7 +135,7 @@ struct TexturedMesh {
 struct Bone {
     Quaternion orientaion;
     Vector3 position;
-    u32 parentIndex;
+    f32 parentIndex;
 };
 
 struct MeshAnimation {
